move more code into DraftPrescriptionsProvider (#2275)
* prescriptionIds

* move all draft prescription CRUD methods

* fix props

* fix tests

* fix provider nesting order

* add error toast

* fix whitespace
